Run clang-format on CUDA and CPP files
Runs clang-format on all CUDA and CPP files in order to improve
readability.
Used the committed .clang-format, which is based off of Google's style
guide with the following changes:

 - IndentWidth: 4
 - AlwaysBreakAfterReturnType: All
 - BinPackArguments: false
 - BinPackParameters: false

To indent with four spaces, put function types on their own line, and to
make function parameters and arguments one per line.